Make NodeID a union type
NodeID is currently a polymorphic type with different implementations
behind an interface. This makes marshalling a bit different for NodeID
since the decoder first has to determine the type and then create the
right object.

By making NodeID a regular struct this problem goes away and NodeID
is created like any other type.