fix tests
github.com/google/go-cmp does not compare unexported fields by default
and there is no straightforward way to enable this. There is a
discussion on https://github.com/google/go-cmp/issues/40 and a suggested
workaround but that still feels clumsy.

During my time at eBay we developed a comparator which handles the usual
test cases fine and which is a small change after the test refactoring
from #128.