Remove reflection from SecureChannel code
This patch removes reflection code in the secure channel.

The Secure Channel implementation relied on reflection to detect service
requests and responses and to get to their respective header fields.

This patch extends the code generator to generate helper methods for
getting and setting the RequestHeader and ResponseHeader fields so that
the reflection code can be removed with type safe code using interfaces.