Issue #313: Decouple dial context from monitor context
This patch decouples the context for the monitor go routine which
dispatches incoming messages to handler functions from the dial context.

The coupling had the effect that the monitor go routine terminates when
the dial context is closed. This can happen even after a successful
connection when a timeout context is used to implement a dial timeout.
The timeout context fires at some later point when it is no longer used
which then terminates the monitor routine and the application hangs.

Fixes #313