browse: do not set timestamp for empty view
To browse the full nodeset the client should use an empty (zero) ViewID
in the browse request. The server should probably ignore the timestamp
and/or version field but ABB 800xA does not return any nodes if the
ViewID is zero but the timestamp field is set.

This patch removes the timestamp field for the empty view.