Check types on creation of Variant #443
The commit add a check to respect the OPC/UA specification.
We refer to Part6 Section 5.1.2 (Table - 1).
This specific reference is being update in this commit as well.

We try to stick as much as possible to the reference without breaking existng tests.

In particular I find unclear if slice and arrays should be supported,
they were supported to I kept backward compatibility.

Moreover I find unclear if the DiagnosticInfo type should be supported.