Refactor type registry to use NodeIDs as keys
This patch changes the TypeRegistry to use NodeIDs as keys.

To fix #444 we need the Lookup method to always return a NodeID.

This was already the case when registering extension objects. However,
the type registry was also used for registering services by service
number in which case only integers were used.