config: bubble up errors from configuration
The current config code has some options which can have errors. The
current code calls `log.Fatal` which terminates the application instead
of returning the error.

The correct solution would be to return the error in ApplyConfig and
then return it in NewClient. However, this would break all existing
clients.

We therefore postpone this breaking change to v0.5.0 when we update the
API with breaking changes anyway and in the meantime bubble the error up
in the Connect and Dial method.

Fixes #616