setSession to nil when recreating to avoid unncessary call to CloseSession.
In the client.go monitor loop, the recreateSession action builds a new session on the assumption that the server is no longer holding the previous. It does not clear the value of atomicSession, which leads to CloseSession being called in ActivateSession even if there are no old sessions to be removed.

This behvior was referenced in issue #673.