subscription: refactor subscription recreation
Decouple client and subscription code during subscription recreation.
During this operation both the state of the client and the subscription
have to be updated while holding a lock in the client.

The previous implementation had both classes call each other. The
control flow in the new implementation is only from the client to
the subscription.